Refactor for performance using lifecycle methods and re structuring folders
