+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from './components/Layout/Layout'
+import Layout from './hoc/Layout/Layout'
 import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder';
 
 function App() {
