+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,29 +1,32 @@
-import React from 'react';
-import Aux from '../../../hoc/Aux';
+import React, {Component} from 'react';
+import Aux from '../../../hoc/Aux/Aux';
 import Button from '../../UI/Button/Button';
 
-const orderSummary = (props) => {
-    const ingredientsSummary = Object.keys(props.ingredients).map((ingredientKey) => {
-        return (
-            <li key={ingredientKey}>
-                <span style={{textTransform: 'capitalize'}}>{ingredientKey}</span>: {props.ingredients[ingredientKey]}
-            </li>
-        );
-    });
+class OrderSummary extends Component {
 
-    return (
-        <Aux>
-            <h3>Your Order</h3>
-            <p>A delicius burger with the following ingredients:</p>
-            <ul>
-                {ingredientsSummary}
-            </ul>
-            <p><strong>Total Price: {props.price.toFixed(2)}</strong></p>
-            <p>Continue to Checkout?</p>
-            <Button buttonType="Danger" clicked={props.cancelPurchase}>CANCEL</Button>
-            <Button buttonType="Success" clicked={props.continuePurchase}>CONTINUE</Button>
-        </Aux>
-    );
+    render() {
+        const ingredientsSummary = Object.keys(this.props.ingredients).map((ingredientKey) => {
+            return (
+                <li key={ingredientKey}>
+                    <span style={{textTransform: 'capitalize'}}>{ingredientKey}</span>: {this.props.ingredients[ingredientKey]}
+                </li>
+            );
+        });
+    
+        return (
+            <Aux>
+                <h3>Your Order</h3>
+                <p>A delicius burger with the following ingredients:</p>
+                <ul>
+                    {ingredientsSummary}
+                </ul>
+                <p><strong>Total Price: {this.props.price.toFixed(2)}</strong></p>
+                <p>Continue to Checkout?</p>
+                <Button buttonType="Danger" clicked={this.props.cancelPurchase}>CANCEL</Button>
+                <Button buttonType="Success" clicked={this.props.continuePurchase}>CONTINUE</Button>
+            </Aux>
+        );
+    }
 }
 
-export default orderSummary;
+export default OrderSummary;
